fix(singletons): fix build issues and move CalendarContext interface back
