+++ packages/sanity/src/_singletons/form/inputs/PortableText/PortableTextMemberItemsContext.ts
@@ -1,4 +1,7 @@
 import {createContext} from 'react'
 import type {PortableTextMemberItem} from 'sanity'
 
+/**
+ * @internal
+ */
 export const PortableTextMemberItemsContext = createContext<PortableTextMemberItem[]>([])
