+++ packages/sanity/src/_singletons/form/studio/PresenceContext.ts
@@ -1,4 +1,7 @@
 import {createContext} from 'react'
 import type {FormNodePresence} from 'sanity'
 
+/**
+ * @internal
+ */
 export const PresenceContext = createContext<FormNodePresence[]>([])
