+++ packages/sanity/src/_singletons/user-color/UserColorManagerContext.ts
@@ -1,4 +1,7 @@
 import {createContext} from 'react'
 import type {UserColorManager} from 'sanity'
 
+/**
+ * @internal
+ */
 export const UserColorManagerContext = createContext<UserColorManager | null>(null)
