+++ packages/sanity/src/core/studio/components/navbar/search/components/filters/filter/inputs/date/datePicker/calendar/contexts/CalendarContext.ts
@@ -1,0 +1,13 @@
+export interface CalendarContextValue {
+  date?: Date
+  endDate?: Date
+  focusedDate: Date
+  selectRange?: boolean
+  selectTime?: boolean
+
+  /**
+   * An integer indicating the first day of the week.
+   * Can be either 1 (Monday) or 7 (Sunday).
+   */
+  firstWeekDay: 1 | 7
+}
